refactor(routes): hoist route registries to shared/, break server->client import
Task 8 Slice 1. server/route-policy/api-route-policy-registry.ts imported
ROUTE_GOVERNANCE_REGISTRY from client/src/app/ -- a server->client dependency.
Converge on a shared manifest DERIVED from the existing registries (relocated
via git mv, content-identical; not rewritten).

- Move app-route-definitions.ts + route-governance-registry.ts to shared/routes/
  (both pure data; app-route-definitions has zero imports).
- Repoint all 8 importers to shared: server policy + verify-route-policy script
  + 3 route-policy tests + the .tsx governance test + the source-guard path
  string + client app-routes.tsx / route-control-flags.ts (@shared/routes/...).

No server/**.ts file imports client/** anymore (grep-clean; remaining matches
are docs/comments).

Gates: npm run check 0-new (3 tsconfigs), 4 repointed tests 29/29, lint clean,
policy:verify PASS (56 entries), build passed.

Follow-ups: Slice 1b forbid-guard (server/** !-> client/**, proven-teeth);
Slice 2 dual-surface request corpus vs makeApp AND registerRoutes.
