+++ scripts/verify-route-policy.ts
@@ -16,7 +16,7 @@ import {
 import {
   ROUTE_GOVERNANCE_REGISTRY,
   type RouteGovernanceEntry,
-} from '../client/src/app/route-governance-registry';
+} from '../shared/routes/route-governance-registry';
 import {
   buildPrototypeFinancialBlockedError,
   type PortfolioPrototypeRouteId,
